Flatten library structure, move example readelf to bin subdirectory
