+++ crates/elf/src/identity.rs
