+++ crates/elf/src/symbol.rs
